+++ src/os/windows/io.rs
@@ -9,7 +9,10 @@ use crate::{
 };
 use windows_sys::Win32::{
     Globalization::CP_UTF8,
-    System::Console::{GetStdHandle, SetConsoleOutputCP, WriteConsoleA, STD_OUTPUT_HANDLE},
+    System::Console::{
+        GetStdHandle, ReadConsoleA, SetConsoleOutputCP, WriteConsoleA, CONSOLE_READCONSOLE_CONTROL,
+        STD_INPUT_HANDLE, STD_OUTPUT_HANDLE,
+    },
 };
 
 /// Initializes the `nstd.os.windows.io` module. This function should be called before any others in
@@ -46,12 +49,42 @@ pub unsafe extern "C" fn nstd_os_windows_io_print_line(cstr: *const NSTDChar) ->
     a | b
 }
 
+/// Retrieves a handle to stdin.
+/// Returns: `NSTDOSWindowsIOHandle stdin` - The standard input stream.
+#[inline]
+#[cfg_attr(feature = "clib", no_mangle)]
+pub unsafe extern "C" fn nstd_os_windows_io_stdin() -> NSTDOSWindowsIOHandle {
+    STD_INPUT_HANDLE
+}
+
 /// Retrieves a handle to stdout.
 /// Returns: `NSTDOSWindowsIOHandle stdout` - The standard output stream.
 #[inline]
 #[cfg_attr(feature = "clib", no_mangle)]
 pub unsafe extern "C" fn nstd_os_windows_io_stdout() -> NSTDOSWindowsIOHandle {
     STD_OUTPUT_HANDLE
+}
+
+/// Reads a buffer from `stream` into `buffer`.
+/// Parameters:
+///     `const NSTDOSWindowsHandle stream` - Handle to an IO stream.
+///     `NSTDSlice *const buffer` - The buffer to read into.
+///     `NSTDUInt32 *const read` - Returns as the number of bytes actually read.
+/// Returns: `NSTDErrorCode errc` - Nonzero on error.
+#[inline]
+#[cfg_attr(feature = "clib", no_mangle)]
+pub unsafe extern "C" fn nstd_os_windows_io_read(
+    stream: NSTDOSWindowsHandle,
+    buffer: &mut NSTDSlice,
+    read: *mut u32,
+) -> NSTDErrorCode {
+    const EOL: CONSOLE_READCONSOLE_CONTROL = CONSOLE_READCONSOLE_CONTROL {
+        nLength: std::mem::size_of::<CONSOLE_READCONSOLE_CONTROL>() as u32,
+        nInitialChars: 0,
+        dwCtrlWakeupMask: 1 << b'\n',
+        dwControlKeyState: 0,
+    };
+    (ReadConsoleA(stream, buffer.ptr.raw, buffer.size as u32, read, &EOL) == 0) as NSTDErrorCode
 }
 
 /// Writes a buffer of `bytes` to `stream`.