+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Latest
+### `nstd.os`
+- Added stdin support to `windows.io`.
 # 0.9.7
 ### `nstd.events`
 - `NSTDEventCallbacks::on_window_[cursor_moved|line_scroll]` now use primitives instead of